- installed testcontainers. - added users service integration and unit tests. - fixed deleting uploaded file image if it exists and user creation fails

+++ src/users/users.service.ts
@@ -97,7 +97,7 @@ export class UsersService {
       return UserMapper.toResponseDTO(savedUser, languages);
     } catch (err) {
       await queryRunner.rollbackTransaction();
-      await deleteFile(profileImage.path);
+      if (profileImage) await deleteFile(profileImage.path);
       throw err;
     } finally {
       await queryRunner.release();
